Print an object

This base method prints objects; here the arguments specific to mclm implementations are described.

Usage

# S3 method for assoc_scores
print(
  x,
  n = 20,
  from = 1,
  freeze_cols = NULL,
  keep_cols = NULL,
  drop_cols = NULL,
  from_col = 1,
  sort_order = c("none", "G_signed", "PMI", "alpha"),
  extra = NULL,
  ...
)

# S3 method for conc
print(x, n = 30, ...)

# S3 method for fnames
print(
  x,
  n = 20,
  from = 1,
  sort_order = c("none", "alpha"),
  extra = NULL,
  hide_path = NULL,
  ...
)

# S3 method for freqlist
print(x, n = 20, from = 1, extra = NULL, ...)

# S3 method for slma
print(x, n = 20, from = 1, ...)

# S3 method for tokens
print(x, n = 20, from = 1, extra = NULL, ...)

# S3 method for types
print(x, n = 20, from = 1, sort_order = c("none", "alpha"), extra = NULL, ...)

Arguments

x

An object of any of the classes for which the method is implemented.

n

Maximum number of items in the object to be printed at once.

from

Index of the first item to be printed.

freeze_cols

Names of columns that should not be affected by the argument from_col. Frozen columns are always printed to the left of non-frozen columns, even if in their original order was different. The names of the types are always and unavoidably printed as the leftmost column.

If this argument is NULL, then the default setting applies, meaning that the following columns, if present, are displayed in the "frozen area": a, PMI and G_signed.

To avoid any columns for being frozen, freeze_cols should be NA or character(0).

keep_cols, drop_cols

A vector of column names or NULL. If both arguments are NULL, all columns are printed (or as many as fit on the screen). If keep_cols is not NULL, it indicates the columns that should be printed. If it is NULL but drop_cols is not, then drop_cols indicates the columns that should not be printed. Note that they have no effect on the frozen area.

Columns that are blocked from printing by these arguments are still available to sort_order.

from_col

Index of the first column to be displayed in the regular area (among all selected columns, including frozen columns). If from_col points

sort_order

Order in which the items are to be printed. In general, possible values are "alpha" (meaning that the items are to be sorted alphabetically), and "none" (meaning that the items are not to be sorted). If x is an object of class assoc_scores, a column name or vector of column names may be provided instead.

extra

Extra settings, as an environment. Arguments defined here take precedence over other arguments. For instance, if extra$from_col is not NULL, it will overrule the from_col argument.

...

Additional printing arguments.

hide_path

A character string with a regular expression or NULL. If it is not NULL, the character string will be removed from the paths when printing.

Value

Invisibly, x. For objects of class assoc_scores, the output consists of two areas: the 'frozen area' on the left and the 'regular area' on the right. Both areas are visually separated by a vertical line (|). The distinction between them is more intuitive in explore(), where the frozen columns do not respond to horizontal movements (with the r and l commands). The equivalent in this method is the from_col argument.