Make all values strictly higher than zero

This is an auxiliary function that makes all values in numeric vector x strictly positive by replacing all values equal to or lower than zero with the values in small.pos. small_pos stands for 'small positive constant'.

Usage

zero_plus(x, small_pos = 1e-05)

Arguments

x

A numeric vector.

small_pos

A (small) positive number to replace negative values and 0s.

Value

A copy of x in which all values equal to or lower than zero are replaced by small_pos.

Examples

(x <- rnorm(30))
#>  [1] -0.39893750 -1.41482664 -1.10020725 -0.18209643  0.61309467 -0.97108318
#>  [7] -0.79663246 -1.55087606  1.31992345  0.60703943  1.17259100 -1.02771757
#> [13]  0.54763021  0.45614215 -0.06761640  0.15320899 -1.01932758 -1.32132570
#> [19] -1.83277199 -1.07435099  0.29274540  0.23111095  0.46413489 -0.99429742
#> [25] -0.25438530  0.77663941  0.38508892 -0.08189568 -0.05018019  0.59679408
zero_plus(x)
#>  [1] 0.0000100 0.0000100 0.0000100 0.0000100 0.6130947 0.0000100 0.0000100
#>  [8] 0.0000100 1.3199234 0.6070394 1.1725910 0.0000100 0.5476302 0.4561422
#> [15] 0.0000100 0.1532090 0.0000100 0.0000100 0.0000100 0.0000100 0.2927454
#> [22] 0.2311109 0.4641349 0.0000100 0.0000100 0.7766394 0.3850889 0.0000100
#> [29] 0.0000100 0.5967941