Make all values strictly higher than zero

This is an auxiliary function that makes all values in numeric vector x strictly positive by replacing all values equal to or lower than zero with the values in small.pos. small_pos stands for 'small positive constant'.

Usage

zero_plus(x, small_pos = 1e-05)

Arguments

x

A numeric vector.

small_pos

A (small) positive number to replace negative values and 0s.

Value

A copy of x in which all values equal to or lower than zero are replaced by small_pos.

Examples

(x <- rnorm(30))
#>  [1]  0.73590420  1.80063821  0.70872217  1.68482245  1.60903903 -0.60443796
#>  [7]  2.30511056  0.67216580 -0.20695340 -0.38364395  0.35348617  0.63177901
#> [13]  1.54573295 -0.57590690 -0.58569503  0.72894925 -0.91698322 -0.37403847
#> [19]  1.09805672  1.00795088  0.52726743 -0.05689228  0.18284248  0.81015510
#> [25]  0.47036815 -1.50922665  0.02228813 -0.22299303 -2.84564781  0.43513545
zero_plus(x)
#>  [1] 0.73590420 1.80063821 0.70872217 1.68482245 1.60903903 0.00001000
#>  [7] 2.30511056 0.67216580 0.00001000 0.00001000 0.35348617 0.63177901
#> [13] 1.54573295 0.00001000 0.00001000 0.72894925 0.00001000 0.00001000
#> [19] 1.09805672 1.00795088 0.52726743 0.00001000 0.18284248 0.81015510
#> [25] 0.47036815 0.00001000 0.02228813 0.00001000 0.00001000 0.43513545