Make all values strictly higher than zero

This is an auxiliary function that makes all values in numeric vector x strictly positive by replacing all values equal to or lower than zero with the values in small.pos. small_pos stands for 'small positive constant'.

Usage

zero_plus(x, small_pos = 1e-05)

Arguments

x

A numeric vector.

small_pos

A (small) positive number to replace negative values and 0s.

Value

A copy of x in which all values equal to or lower than zero are replaced by small_pos.

Examples

(x <- rnorm(30))
#>  [1] -1.27557989  0.34385416 -1.23519715 -0.65830154  1.39673530  0.23382965
#>  [7]  0.60721547  0.89569540  0.93851316 -0.73339665  2.46214767  1.07063450
#> [13]  1.05579262  1.82148935 -0.82896834  0.12519205  0.23553131  2.07138764
#> [19] -0.50776218  0.43313599 -2.29952617 -1.43858901 -1.76791221  0.78644712
#> [25]  1.27813801  0.16730794 -0.66433718 -0.02999350 -0.01641384  0.88203493
zero_plus(x)
#>  [1] 0.0000100 0.3438542 0.0000100 0.0000100 1.3967353 0.2338296 0.6072155
#>  [8] 0.8956954 0.9385132 0.0000100 2.4621477 1.0706345 1.0557926 1.8214894
#> [15] 0.0000100 0.1251921 0.2355313 2.0713876 0.0000100 0.4331360 0.0000100
#> [22] 0.0000100 0.0000100 0.7864471 1.2781380 0.1673079 0.0000100 0.0000100
#> [29] 0.0000100 0.8820349